requirements.txt
Django==5.0.6
cryptography==42.0.7
jsons==1.6.3
requests==2.32.3
shortuuid==1.0.13
django-allauth==0.63.3
json5==0.9.25
jsonpointer==3.0.0
jsons==1.6.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pillow==10.3.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
webencodings==0.5.1
websocket-client==1.8.0
PyJWT==2.8.0
psycopg2==2.9.9
